requirements.txt
tensorflow
opencv-python
pandas
numpy
h5py
pyserial
